requirements.txt
appdirs==1.4.3
Django==1.10.6
geopy==1.11.0
mysqlclient==1.3.10
packaging==16.8
progressbar2==3.12.0
pyparsing==2.2.0
python-utils==2.0.1
six==1.10.0
